Add main menu sub menu for categories of unlockables
