+++ Assets/Scripts/UI/DescriptionMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26b61514aa6c29244bdfb8a8cbb15041
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
